Separe el codigo en headers
