+++ maze.h
@@ -1,0 +1,29 @@
+#include <stdio.h>
+
+#define TAM 10
+
+int maze[TAM][TAM] = {
+    {0, 0, 0, 1, 1, 1, 1, 1, 1, 1},
+    {1, 1, 0, 0, 0, 0, 0, 0, 0, 1},
+    {1, 1, 0, 1, 0, 1, 0, 1, 0, 1},
+    {0, 0, 0, 1, 1, 1, 0, 1, 0, 1},
+    {0, 1, 0, 0, 0, 0, 1, 0, 0, 0},
+    {0, 1, 0, 1, 1, 0, 1, 0, 1, 0},
+    {0, 0, 1, 1, 1, 0, 1, 0, 1, 0},
+    {1, 0, 0, 0, 0, 1, 1, 0, 1, 0},
+    {0, 0, 1, 1, 1, 1, 1, 1, 1, 1},
+    {0, 0, 0, 0, 0, 0, 0, 0, 0, 0}};
+
+int playMaze[TAM][TAM] = {
+    {8, 0, 0, 1, 1, 1, 1, 1, 1, 1},
+    {1, 1, 0, 0, 0, 0, 0, 0, 0, 1},
+    {1, 1, 0, 1, 0, 1, 0, 1, 0, 1},
+    {0, 0, 0, 1, 1, 1, 0, 1, 0, 1},
+    {0, 1, 0, 0, 0, 0, 1, 0, 0, 0},
+    {0, 1, 0, 1, 1, 0, 1, 0, 1, 0},
+    {0, 0, 1, 1, 1, 0, 1, 0, 1, 0},
+    {1, 0, 0, 0, 0, 1, 1, 0, 1, 0},
+    {0, 0, 1, 1, 1, 1, 1, 1, 1, 1},
+    {0, 0, 0, 0, 0, 0, 0, 0, 0, 0}};
+
+int solucion[TAM][TAM];
